Исправлено

+++ src/main/java/ru/kata/spring/boot_security/demo/repository/UserRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.kata.spring.boot_security.demo.model.User;
 
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-
+public interface UserRepository extends JpaRepository<User, Long> {
     User findByLogin(String login);
-
 }
